# optimize u128/i128 popcounts further