Don't fall back on breaking apart the popcount operation if 128-bit integers are natively supported.