@antoyo 2023-10-18
Thanks!

Do you think it would be worth trying to make it work even for non-native integers by using the operations from the `int` module as I suggested earlier?