@sadlerap 2023-10-18
I'll give it a try - if it ends up working, I'll replace what I've posted here with whatever works.  If not, I'll let you know what fails.