@antoyo 2023-10-18
Indeed. Good find!
You can see in the [CI](https://github.com/sadlerap/rustc_codegen_gcc/actions/runs/6555101043/job/17803082877#step:20:4317) that it's because you use an array for a binary operation.
It's in [the call to `add_assignment_op()`](https://github.com/sadlerap/rustc_codegen_gcc/commit/c715c88d06335c88f74b63060aa676766eb0b508#diff-34f8c479ac3b1727ae8bccaf335c8f3be407f4d48dbf725e26b793ebee763dc4R849).
libgccjit doesn't check that the values in this call are numeric.
I'll fix this!
Thanks!