@sadlerap 2023-10-18
hmm, I wonder if replacing the assignment ops with just plain assignments would work.  I'll give it a try.