@antoyo 2023-10-19
I added the type check in [this commit](https://github.com/antoyo/gcc/commit/2eca081142f9161adda821812f74c6669935fe0a).

Yeah, that might work indeed.