@sadlerap 2023-10-19
Sadly, while subbing out assignment ops with pure assignments works around the bug, it also results in poor codegen when 128-bit integers aren't supported and [the fallback][1] is skipped.

[1]: https://github.com/rust-lang/rustc_codegen_gcc/blob/a9f31ab4814a936c91dec8be0da8807c7210b44a/src/intrinsic/mod.rs#L822-L831

<details><summary>Details</summary>
<p>


Before:

```asm
popcount_128:

        xor eax, eax
        popcnt rax, rsi
        test rsi, rsi
        cmove rax, rsi

        xor ecx, ecx
        popcnt rcx, rdi
        mov rdx, rax
        add rax, rcx

        test rdi, rdi
        cmove rax, rdx

        ret
```

After:

```asm
popcount_128:

        push r14

        mov rax, rsi
        push r13
        push r12
        push rbp
        push rbx
        sub rsp, 32

        or rax, rdi
        mov QWORD PTR 16[rsp], rdi

        mov QWORD PTR 24[rsp], rsi
        vmovdqa xmm3, XMMWORD PTR 16[rsp]
        vmovdqa XMMWORD PTR [rsp], xmm3
        je .L67
        mov rbp, rsi

        mov rbx, rdi

        xor r12d, r12d

.L66:
        mov r14, QWORD PTR [rsp]
        mov r13, QWORD PTR 8[rsp]
        xor ecx, ecx
        mov edx, 1
        add r12d, 1
        mov rdi, r14
        mov rsi, r13
        call __rust_u128_sub@PLT

        xor ecx, ecx
        mov edx, 1
        mov rdi, r14
        mov rsi, r13
        and rbx, rax
        mov QWORD PTR 16[rsp], rax
        call __rust_u128_sub@PLT

        vmovq xmm0, QWORD PTR 16[rsp]
        mov rax, rbx
        and rbp, rdx
        vpinsrq xmm0, xmm0, rdx, 1
        vpand xmm1, xmm0, XMMWORD PTR [rsp]
        or rax, rbp
        vmovdqa XMMWORD PTR [rsp], xmm1
        jne .L66
        mov eax, r12d
.L63:
        add rsp, 32
        pop rbx

        pop rbp

        pop r12

        pop r13
        pop r14

        ret
.L67:

        xor eax, eax
        jmp .L63
```

</p>
</details> 

However, when 128-bit integers are supported, we do see the improvement as expected.

<details><summary>Details</summary>
<p>

Before: 

```asm
popcount_128:

        xor eax, eax
        popcnt rax, rsi
        test rsi, rsi
        cmove rax, rsi

        xor ecx, ecx
        popcnt rcx, rdi
        mov rdx, rax
        add rax, rcx

        test rdi, rdi
        cmove rax, rdx

        ret
```

After:

```asm
popcount_128:

        mov rdx, rdi
        or rdx, rsi

        je .L65

        xor ecx, ecx

        popcnt rsi, rsi

        popcnt rcx, rdi
        lea eax, [rsi+rcx]
        ret
.L65:

        xor eax, eax
        ret
```

</p>
</details> 

I think if we gate the fallback behind native support for 128-bit integers, we'll get the best of both worlds - good-ish codegen when 128-bit integers are supported, and a roughly as-good fallback when they aren't.