@antoyo 2023-10-19
Do you know (or did you do some benchmarks to check) if having one jump is indeed faster than the 2 conditional moves?