@sadlerap 2023-10-19
I wrote some microbenchmarks with `criterion`, but they seemed to get stuck on the anaylsis step (let it run for about 15 minutes and it still wasn't done, whereas the same code compiled with llvm completed in a few seconds).  I'll try some other options, see if I get more luck there.